Use GroupsCsvGeneratorService for GitLab.
